test: moved test to messaging from SDK
